Clean up unused comment codes, Change "GetTollFee" method return type from double to decimal, add new method "CheckIfDayBeforeHoliday" and return true for free fee for the day before the public holiday day ,Test the method.

+++ GbgRoadTollSystemTest/FeePerDayOperationTests.cs
@@ -24,9 +24,9 @@ namespace RoadTollSystemTest
             var thirdResult3 = feePerDayOperation.GetFeeInCertainTime(thirdTimes);
 
             //Actual fee in certain time
-            double feeFirstTime = 9;
-            double feeSecondTimes = 22;
-            double feeThirdTime = 16;
+            decimal feeFirstTime = 9;
+            decimal feeSecondTimes = 22;
+            decimal feeThirdTime = 16;
 
             Assert.Equal(feeFirstTime, firstResult);
             Assert.Equal(feeSecondTimes, secondResult2);
